fix： 修复了FDID转换失败的bug

+++ src/main/java/com/example/family/user/dto/UserInfoDto2.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 @Data
 public class UserInfoDto2 {
 
-    private Integer userFdId;
+    private Long userFdId;
 
     private String userName;
